241108 상품 이미지 업로드 안됨 이미지 제외하고 상품등록은 가능

+++ src/main/java/com/example/byosinsa/repository/ProductImgRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductImgRepository extends JpaRepository<ProductImg, Long> {
+public interface ProductImgRepository extends JpaRepository<ProductImg, Integer> {
 
     List<ProductImg> findByProduct_Pno(int pno);
     //aa
